requirements.txt
# python==3.8.2 https://www.python.org/downloads/windows/ -> Download: Windows x86-64 executable installer
# setup: pip install -r requirements.txt
# run: python manage.py runserver
# ----------------------------------------------
django==3.0.4
djangorestframework==3.11.0
django-cors-headers==3.2.1
PyMySQL==0.9.3
PyJWT==1.7.1
gunicorn==20.0.4